moved module classes to layer classes, deleted modules package
