--- src/Hardware/Sensors/Button/Button2
@@ -1,31 +1,0 @@
-package Hardware.Sensors.Button;
-
-import ButterCat.HelpFunctions;
-
-public class Button2
-{
-    private int pin;
-    private boolean isPressed;
-
-    public Button2(int pin) {
-        HelpFunctions.checkDigitalPin("Button pin", pin);
-
-        this.pin = pin;
-        this.isPressed = false;
-        onButtonPress = () -> callback();
-    }
-
-    private callback() {
-        // Nothing here
-    }
-
-    public ButtonCallback onButtonPress;
-
-    public int getPin() { return pin; }
-    public void setPin(int pin) { this.pin = pin; }
-
-    public boolean isPressed() { return isPressed; }
-    public void setPressed(boolean pressed) { isPressed = pressed; }
-
-
-}
