--- src/Hardware/Sensors/Button/ButtonCallback
@@ -1,5 +1,0 @@
-package Hardware.Sensors.Button;
-
-public interface ButtonCallback {
-    void onButtonPress(Button2 source);
-}
