--- src/Hardware/Sensors/Led/LedInterface.java
@@ -1,8 +1,0 @@
-package Hardware.Sensors.Led;
-
-public interface LedInterface {
-    void turnOn();
-    void turnOff();
-    void blink(double frequency);
-    boolean getStatus();
-}
