--- src/buttercat/modules/Callback.java
@@ -1,6 +1,0 @@
-package buttercat.modules;
-
-public interface Callback
-{
-    void run();
-}
